Guia 07 Collections Ejercicio 04

+++ Guia 07 Java Colecciones/guia07_ej_04/src/guia07_ej_04/entidades/Pelicula.java
@@ -17,18 +17,18 @@ public class Pelicula {
 
     private String titulo;
     private String director;
-    private int duracionHs;
+    private double duracionHs;
 
     public Pelicula() {
     }
 
-    public Pelicula(String titulo, String director, int duracionHs) {
+    public Pelicula(String titulo, String director, double duracionHs) {
         this.titulo = titulo;
         this.director = director;
         this.duracionHs = duracionHs;
     }
 
-    public int getDuracionHs() {
+    public double getDuracionHs() {
         return duracionHs;
     }
